MVP-5447-3: Refactor and modulize helpers
