+++ packages/notifi-web-push-service-worker/lib/index.ts
@@ -1,5 +1,20 @@
 import { NotifiEnvironment } from '@notifi-network/notifi-frontend-client';
 
+export * from './notifi-service';
+export * from './utils';
+
+export type NotifiServiceWorkerMessagePayload = {
+  type: NotifiServiceWorkerMessageType;
+  userAccount: string;
+  dappId: string;
+  env: NotifiEnvironment;
+};
+
+export enum NotifiServiceWorkerMessageType {
+  NotifiCheckSubscription = 'NotifiCheckSubscription',
+}
+
+// TODO: Modulize this function when growing big (maybe utils/service-worker.ts)
 export function initWebPushServiceWorker(
   serviceWorkerFilePath?: string | null,
   userAccount?: string | null,
@@ -18,7 +33,7 @@ export function initWebPushServiceWorker(
         }
 
         const payload = {
-          type: 'NotifiCheckSubscription',
+          type: NotifiServiceWorkerMessageType.NotifiCheckSubscription,
           userAccount: userAccount,
           dappId: dappId,
           env: env,
@@ -33,19 +48,20 @@ export function initWebPushServiceWorker(
   }
 }
 
+// TODO: Modulize this function when growing big (maybe utils/service-worker.ts)
 export function tryCreateWebPushSubscription(
-  userAccount: string,
+  userAccount: string, // IMPORTANT: Case sensitive, make sure to pass the account identically as it is in (web) app
   dappId: string,
   env: NotifiEnvironment,
 ) {
   console.log('tryCreateWebPushSubscription was called');
   if ('serviceWorker' in navigator) {
     navigator.serviceWorker.getRegistration().then((registration) => {
-      const payload = {
-        type: 'NotifiCheckSubscription',
-        userAccount: userAccount,
-        dappId: dappId,
-        env: env,
+      const payload: NotifiServiceWorkerMessagePayload = {
+        type: NotifiServiceWorkerMessageType.NotifiCheckSubscription,
+        userAccount,
+        dappId,
+        env,
       };
       registration?.active?.postMessage(JSON.stringify(payload));
     });
@@ -55,3 +71,15 @@ export function tryCreateWebPushSubscription(
     );
   }
 }
+
+export const isNotifiServiceWorkerMessage = (
+  data: unknown,
+): data is NotifiServiceWorkerMessagePayload => {
+  if (typeof data !== 'object') return false;
+  const messagePayload = data as NotifiServiceWorkerMessagePayload;
+  if (!('type' in messagePayload)) return false;
+  if (!('userAccount' in messagePayload)) return false;
+  if (!('dappId' in messagePayload)) return false;
+  if (!('env' in messagePayload)) return false;
+  return true;
+};
