+++ packages/notifi-web-push-service-worker/lib/notifi-service/index.ts
@@ -1,0 +1,1 @@
+export * from './webpush-target';
