+++ packages/notifi-web-push-service-worker/lib/utils/constants.ts
@@ -1,0 +1,5 @@
+export const webPushTargetIdKey: IDBValidKey = 'webPushTargetId';
+
+export const notifiDb = 'notifi';
+
+export const notifiDefaultDbStore = 'keyvaluepairs';
