+++ packages/notifi-web-push-service-worker/lib/utils/conversion.ts
@@ -1,0 +1,16 @@
+export function urlBase64ToUint8Array(base64String: string) {
+  const padding = '='.repeat((4 - (base64String.length % 4)) % 4);
+  const base64 = (base64String + padding).replace(/-/g, '+').replace(/_/g, '/');
+
+  const rawData = self.atob(base64);
+  const outputArray = new Uint8Array(rawData.length);
+
+  for (let i = 0; i < rawData.length; ++i) {
+    outputArray[i] = rawData.charCodeAt(i);
+  }
+  return outputArray;
+}
+
+export function uint8ArrayToBase64Url(uint8Array: Uint8Array) {
+  return self.btoa(String.fromCharCode.apply(null, Array.from(uint8Array)));
+}
